Added rest of neccessary files for the assignment (currently blank). Added information to the HPP file Adjusted for boolean values Comments added to ensure readability of Box.hpp Typed up and readded information for the boolean values Added the main constructor value for when the user adds in values at creation of the object Added in extra comments to the application for better readability Added comments in for every function as per code guidelines Further adjusted the application information.  (last thing needed was adjustments to the boolean values).
Adding personal test to see how well the class works
Adding visual test for console to make sure it is saving values correctly
Completed box Testing
Altered boolean values to be void values because it was not needed to return boolean values.
Moved over the blank constructor to the .cpp file to maintain cleanliness in the .hpp file.
Fixed basic grammar and spelling issues

Re did commenting to include names of functions for better readability